Add algorithm logic
This update includes the team strength computation and the distribution of  participants without a team.

+++ src/main/java/com/jbhunt/infrastructure/universityhackathon/entity/Team.java
@@ -49,9 +49,8 @@ public class Team {
     @Column(name="TeamIconCode")
     private String teamIconCode;
 
-    //TODO: add strength attribute
     //TODO: consider if this attribute should be added to the DB
     @Transient
-    private Integer teamStrength;
+    private Double teamStrength;
 
 }
